Move some tests
